refactor: remove app
