+++ src/routes/index.tsx
@@ -1,12 +1,13 @@
 import React from 'react';
 import { Route, Switch, BrowserRouter } from 'react-router-dom';
 import AppWrapper from '../components/wrappers/AppWrapper';
-import App from '../components/App';
+import Map from '../components/map';
 import ProviderRoutes from '../components/navigation/ProviderRoutes';
 import SentryWrapper from '../components/wrappers/SentryWrapper';
 import AboutUs from '../components/about';
 import Home from '../components/home';
 import { homeRoute, aboutRoute } from './pathnames';
+import 'bootstrap/dist/css/bootstrap.min.css';
 
 function renderComponentWithErrorBoundary<P>(Component: React.ReactType) {
   return (props: P) => (
@@ -22,7 +23,7 @@ const routes = (allRoutes: string[]) => (
       <Switch>
         <Route exact path={homeRoute} component={renderComponentWithErrorBoundary(Home)} />
         {allRoutes.map((route: string) => (
-          <Route path={route} component={renderComponentWithErrorBoundary(App)} />
+          <Route path={route} component={renderComponentWithErrorBoundary(Map)} />
         ))}
         <Route path={aboutRoute} component={renderComponentWithErrorBoundary(AboutUs)} />
         <ProviderRoutes />
